Added tcp check and isolated main logic
